fix status bug and make easier to read
