+++ src/core/useDepositEth/index.ts
@@ -1,1 +1,2 @@
 export * from './useDepositEth';
+export * from './types';
