+++ src/core/useDepositEth/types.tsx
@@ -1,0 +1,7 @@
+export enum DepositEthStatus {
+  IDLE = 'idle',
+  WAITING_FOR_SIGNER = 'waitingForSigner',
+  CONFIRMING = 'confirming',
+  SUCCESS = 'success',
+  ERROR = 'error',
+}
