+++ src/core/useDepositEth/useDepositEth.tsx
@@ -1,60 +1,36 @@
 import { Client } from '@aragon/sdk-client';
-import { DaoDepositSteps, DepositParams } from '@aragon/sdk-client';
-import { useReducer } from 'react';
+import { DepositParams } from '@aragon/sdk-client';
+import { useState } from 'react';
 import { useMutation, UseMutationOptions } from 'react-query';
 
 import { useAragon } from '../../context';
-
-export enum DepositEthStatus {
-  IDLE = 'idle',
-  DEPOSITING = 'depositing',
-  CONFIRMING = 'confirming',
-  SUCCESS = 'success',
-  ERROR = 'error',
-}
-
-const initialState: DepositState = {
-  depositTxid: null,
-  depositStatus: DepositEthStatus.IDLE,
-};
-
-interface DepositState {
-  depositTxid: string | null;
-  depositStatus: DepositEthStatus;
-}
-
-function reducer(state: DepositState, update: Partial<DepositState>) {
-  return { ...state, ...update };
-}
+import { DepositEthStatus } from './types';
 
 export function useDepositEth(depositParams: DepositParams, options?: UseMutationOptions) {
-  const [state, dispatch] = useReducer(reducer, initialState);
+  const [depositTxid, setDepositTxid] = useState<string | null>(null);
+  const [depositStatus, setDepositStatus] = useState<DepositEthStatus>(DepositEthStatus.IDLE);
+
   const { baseClient: client } = useAragon();
 
   const deposit = async (client: Client, depositParams: DepositParams) => {
-    const steps = client.methods.deposit(depositParams);
-    dispatch({ depositStatus: DepositEthStatus.DEPOSITING });
-    for await (const step of steps) {
-      try {
-        switch (step.key) {
-          case DaoDepositSteps.DEPOSITING:
-            dispatch({ depositTxid: step.txHash, depositStatus: DepositEthStatus.CONFIRMING });
-            // console.log({ state, step });
-            break;
-          case DaoDepositSteps.DONE:
-            dispatch({ depositStatus: DepositEthStatus.SUCCESS });
-            // console.log({ state, step });
-            break;
-        }
-      } catch (err) {
-        dispatch({ depositStatus: DepositEthStatus.ERROR });
-        console.error(err);
-      }
+    try {
+      const steps = client.methods.deposit(depositParams);
+      setDepositStatus(DepositEthStatus.WAITING_FOR_SIGNER);
+
+      setDepositTxid((await steps?.next()).value.txHash);
+      setDepositStatus(DepositEthStatus.CONFIRMING);
+
+      await steps?.next();
+      setDepositStatus(DepositEthStatus.SUCCESS);
+    } catch (err) {
+      setDepositStatus(DepositEthStatus.ERROR);
+      console.error(err);
     }
   };
 
   return {
-    ...state,
+    depositTxid,
+    depositStatus,
     ...useMutation({
       mutationKey: ['depositEth', depositParams.daoAddressOrEns],
       mutationFn: async () => await deposit(client, depositParams),
